Update patch set 1
Patch Set 1: -Code-Review

No, the reason it seems like more effort is the idea that this stuff could change. We're dealing with different variants where the the firmware is definitely different between them. If you're comfortable with the notion that these wont change, then I'm fine with it. My only concern is down the road if one variant changes we'll be reverting and going back to the init script method anyway.

Patch-set: 1