Update patch set 1
Patch Set 1:

When it comes down to it, even with the Amps used, the sound going to the speaker is not optimized to be within its targeted efficiency range.  This is why HTC adds effects to the audio that then gets sent to the amp. If you ask an audio expert, of which I have consulted two of, they will all tell you this.  Compression and bass reduction are both used by Boomsound, and this is why HTC does not allow Boomsound to be disabled.

Patch-set: 1