Update patch set 1
Patch Set 1:

At that point it would not make sense to actually even include DSPManager.  It would almost be a requirement to just use hexagon to create a new DSP manager for Snapdragon devices.  However, we are not at that point and it is unknown how long out this is. So as a temporary fix, I still stand behind having a default for the roms to contain until amp is fixed and stable.

Patch-set: 1
Label: Code-Review=0