Update patch set 4
Patch Set 4: Code-Review+1

Runs a bit better; benchmarks quite a bit better. 

Best was 8,000 in Quadrant before; now 10,000; best was 29,000 in AnTuTu before; now 35,000.

I tried AnTuTu X to be sure there's no app spy and got the same results as with the normal app.  I think the app spy is in the pnp xml, which I removed for a couple reasons (couldn't build with it because it's obfuscated, and even adding it manually had performance and benchmark degradations).

Patch-set: 4