Update patch set 1
Patch Set 1:

As far as I can tell this is the case.

I haven't tested this yet fwiw, but EXIT_PROP_NAME never gets set to 1 as far as I can tell.

Only other devices that use this prop in init I have sync'd are d2lte and jflte, I haven't looked into if its working on them or not.

Patch-set: 1