Update patch set 1
Patch Set 1:

The CPU Boost hint commit here https://github.com/CyanogenMod/android_device_qcom_common/commit/5e84777ad1346c4664a32e639af75c4e44d08cd4

Seems to be the cause of it, once this is reverted the cpu min is set properly and the cores are not stuck on max. Tested on the m8, and falcon so far.

Patch-set: 1