Update patch set 22
Patch Set 22:

Nice!

I just ran a test-build for my m8. At first, I missed the two changes referenced in Patch Set 2. In that case it built, but it didn't do anything. When I noticed what happened, I grabbed those as well, and everything worked smoothly.

I haven't tried answering a phone call or anything, but I've seen it while on the charger with a Hangouts notification. I also turned the torch on and off. All works well :)

The dot-view case was a show-off thing I liked when I first got the phone, but the CyanogenMod "profiles" piece was more important to me than that. I wish I could keep this build, but there's something currently "broken"; any app that tries to access the camera crashes. It's not related to these changes though. It's just something that keeps coming and going for me, even in the nightlies.

Anyway... I had a quick glance at the code, but I've been too out-of-touch with it lately to provide any relevant feedback. I just figured a "works for me" and "I'd keep this if I could" comment might help.

Great job! :)

Patch-set: 22