Update patch set 1
Patch Set 1:

Still testing this one, but so far it seems better. Usually I can get the brightness to "stick" at max by switching from Dotcase to the launcher a few times, but so far I have not been able to. I'll use it for a while and see if it ends up sticking at some point again.

Patch-set: 1