Update patch set 1
Patch Set 1:

Details: If Dotcase tries to unregister a receiver that is already unregistered (or hasn't been registered yet), it returns an (unchecked) IllegalArgumentException. This can happen if the activity tries to shut down before it gets started the first time (e.g., in certain cases when case status changes directly after boot).

Catch the exception, log the error, and move on, since it's harmless.

Patch-set: 1