Update patch set 2
Patch Set 2:

Okay, working on modifying how notifications are handled so they are painted in the order the notification manager returns them to us (which I assume is the order they appear in the pulldown?). Also, it will allow us to add more than 6 sprites, and if there are more notifications than we can fit on the screen, we can add something like a "..." sprite.

Patch-set: 2