Update patch set 2
Patch Set 2: -Code-Review

On average, I'd say Dotcase previously crashed once every 36 hours during my normal usage. 36 hours on with this patch, hasn't crashed yet. Assuming it's a random variable, and assuming my probability isn't super rusty (it is), I can guess this is 50% likely to have fixed the problem (and so far, no ill effects).

In another day or so, I'll assume this is at worst equivalent to the old way, and at best a fix for the crash, and +1 it accordingly (at which point I'll pass the torch to Dan for his judgment).

Patch-set: 2
Label: Code-Review=0