Update patch set 2
Patch Set 2:

hammerhead config is based on Marc vdW's interactive patches, one of which is causing excessive CPU drain.  You may want to take settings from before this commit:

commit 9e4e812f46f9fff09833d05e3b002bd2a5a7ca10
Author: myfluxi <linflux@arcor.de>
Date:   Sat Jul 19 00:43:25 2014 +0200

    hammerhead: Reconfigure sync freq feature

    Now that interactive is working properly, there is no
    need for overly aggressive cpu freqs. Also, remove
    the unused config for target loads.

    Change-Id: I36e2fecfc24f0dd7da8224638a902d2c3be83b47

At the very last, you may want to think about whether or not you want to add

    write /sys/devices/system/cpu/cpufreq/interactive/target_loads "85 1500000:90 1800000:70"

which was removed because his patchset removes this tunable.   The normal default for this is just "85".

Patch-set: 2