Update patch set 1
Patch Set 1: Code-Review-1

This is not ready. Lots of spacing issues, several characters are impossible to read (N, W, etc), and it needs a clean-up.

Also, we'll probably have to go variable-width for this; we just can't fit a reasonable-looking W in only 3 pixels, and if we go wider but stay fixed-width, barely anything fits on the screen.

Patch-set: 1