Update patch set 2
Patch Set 2:

No, not really, I am for it I think when it comes down to it.  I don't use the torch very much from the dotcase and when I do it doesn't register a long press and I am flipping the case open to turn on the torch. Other times it comes on at the wrong time. I would leave the option to turn it off from the case though. Just my .02 cents.

Patch-set: 2