Update patch set 1
Patch Set 1: Code-Review-1

Or just torch this file, bring in bacon config, build libqcompostprocbundle, and add libqcbassboost, libqcvirt, and libqcreverb from bacon.

Those files aren't device-specific since not even manufacturers have source for it (if you look at the strings, the path to the file has no-ship in it, meaning that the source is not given as part of the BSP).

Patch-set: 1
Label: Code-Review=-1