Update patch set 1
Patch Set 1:

To be clear: I get the issue steve mentions even with all firmware, radio, bootloader etc. updated to the ones in the GPE 5.0.1 RUU.

I have not tried with the new kernel though as my build including that failed to reach the bootanimation. It somehow corrupted my partitions so I had to format /data and restore a month old backup. I'm doing a fresh clobbered build now.

Patch-set: 1