Update patch set 1
Patch Set 1:

Actually, in the patch you linked, this path is not changed. I was looking at voice-speaker-mic, which I think is the speakerphone mic. I think that one is probably correct as speaker-dmic-endfire (2nd-mic).

This path (handset-mic) is not changed in the patch you linked, it was changed in this one: https://github.com/CyanogenMod/android_device_htc_m8/commit/5dfc8cdca5d7d47dbe851c8eeaa1458c51f0beec

Before that patch, it was set to dmic1, which is what the stock HTC mixer_paths sets. However, dmic1 doesn't work in CM, and everywhere else that stock HTC uses dmic1, we use line-vr.

Patch-set: 1