Update patch set 1
Patch Set 1:

This would make it so that listing individual props wouldn't be required, but unwanted props would be included (inadvertently?). Having the blob files listed one-by-one allows for more control, but it is time consuming and error-prone. If this were implemented, one would think that only wanted blob files would be put into the directory...could this cause an issue in which files that are currently in said directories are unwanted?

Looking as an uneducated outsider, this seems like just another way to skin the cat. Pros and cons to either side.

Patch-set: 1
Reviewer: Alex Elkins <2992@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0