Update patch set 3
Patch Set 3:

Cleans up settings to match stock. It will also make it easier to track changes to the stock firmware on Sprint.

When looking at the two firmwares I found a ton of settings that Verizon uses that Sprint does not. In fact there may be more still that need moving/removing. The end goal is to better match stocks settings

Patch-set: 3