Update patch set 6
Patch Set 6:

I can easily set it to just sprint,  but under request I moved it back to all.  Either way it is just a one line call in init_m8 to set it one way or another.  Could even have an extra class for GSM and vzw if needed

Patch-set: 6