Update patch set 6
Patch Set 6:

https://github.com/CyanogenMod/android_kernel_htc_msm8974/blob/cm-12.0/arch/arm/boot/dts/msm8974.dtsi#L1987

My thought is we could specify some other names in the kernel and if we fail to load one, try others. Not sure that's the best approach either though. Maybe I'll take a quick swing at it.

Also - I uploaded a patch to use install/bin/ (untested as my build box is tied up at the moment)

Patch-set: 6