Update patch set 2
Patch Set 2:

It loads properties on boot or toggle of airplane mode, but it only does it once.  I had the original issue of insane data usage (gigabytes worth) with these blobs, but it's not doing that anymore. I've been running these blobs and this gps.conf for 3 days now.

Patch-set: 2