Update patch set 2
Patch Set 2:

When the blobs were originally updated to 6.0, it was seeing a SIM change or something. Using the GPE blobs with the new gps.conf (without xtra_servers) fixed the data issue, but it takes 5 minutes to get a gps lock. The vzw blobs and vzw gps.conf get an almost immediate lock and it's not constantly seeing a SIM change.

Patch-set: 2